refactor(frontend): вынесена логика задач в отдельные файлы TaskItem, TaskList и TaskForm
