+++ frontend/index.html
@@ -9,3 +9,4 @@
     <script type="module" src="src/main.js"></script>
   </body>
 </html>
+<!-- ToDo если текста слишком много он не влизает в задачу - сделать перенос -->
